+++ package.json
@@ -3,6 +3,7 @@
   "version": "4.6.1",
   "description": "A simple extension to native elements",
   "main": "index.js",
+  "browser": "browser.js",
   "scripts": {
     "start": "wzrd test/index.js:bundle.js",
     "test": "standard && node test/server.js && browserify test/index.js | tape-run",
@@ -28,9 +29,9 @@
   },
   "homepage": "https://github.com/shama/bel",
   "dependencies": {
-    "global": "^4.3.0",
     "hyperx": "^2.3.0",
     "on-load": "^3.2.0",
+    "pelo": "0.0.1",
     "tape-run": "^3.0.0"
   },
   "devDependencies": {
